requirements.txt
cycler==0.10.0
fcsparser==0.2.1
fcswrite==0.6.1
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.1
pandas==1.3.1
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.16.0
sklearn==0.0
threadpoolctl==2.2.0
